// src/actions/types.ts
export enum types {
  MOVE_TO_FRAME = 'MOVE_TO_FRAME',
  SET_POINT_CLASS = 'SET_POINT_CLASS',
  SELECT_TOOL = 'SELECT_TOOL',
  SET_FOCUS_TO_VIEW = 'SET_FOCUS_TO_VIEW',

  UPDATE_FRAME_TYPE = 'UPDATE_FRAME_TYPE',
  UPDATE_DATAPOINT = 'UPDATE_DATAPOINT',
  DELETE_DATAPOINT = 'DELETE_DATAPOINT',
  LOADED_DATASET = 'LOADED_DATASET',
  COPY_FROM_FRAME = 'COPY_FROM_FRAME',
  UNDO = 'UNDO',
  REDO = 'REDO',

  UPDATE_CALCULATED_POINTS = 'UPDATE_CALCULATED_POINTS',
  UPDATE_CALCULATED_POINTS3D = 'UPDATE_CALCULATED_POINTS3D',

  UPDATE_CAM_PARAMS = 'UPDATE_CAM_PARAMS',
  UPDATE_BONE_CONNECTIONS = 'UPDATE_BONE_CONNECTIONS',
  SET_CALCULATION_AVAILABILITY = 'SET_CALCULATION_AVAILABILITY',
  SET_POINT_VISIBILITY = 'SET_POINT_VISIBILITY'
}

export enum ToolSelected {
  SelectMode,
  AddingMode
}

export default types;